+++ tests/Unit/Reference/TagReference/Resolve/Main.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Di\Tests\Unit\Reference\TagReference\Resolve;
+
+final class Main
+{
+    public array $data = [];
+}